Refactor SVG writer class to make multiprocessing easier in Blender
